formatted files

+++ src/app/modals/deleteUser/confirmDelete.component.ts
@@ -16,19 +16,36 @@ export class ConfirmDeleteComponent
 
   constructor(private usersService: UsersService) { }
 
+  //------------Properties------------//
+
+  /**
+   * Gets the first name of the selected user from the users service
+   */
   get userName(): string {
     return this.usersService.selectedUserName.firstName
   }
 
+  /**
+   * Gets the id of the selected user from the users service
+   */
   get userId(): number {
     return this.usersService.selectedUserId
   }
 
-  confirmDelete() {
+  //-----------Methods---------------//
+  /**
+   * Sends user id to users service to be deleted
+   */
+  confirmDelete() 
+  {
     this.usersService.deleteUser(this.userId)
   }
 
-  cancel() {
+  /**
+   * Closes the delete modal
+   */
+  cancel() 
+  {
     this.usersService.showConfirmDelete = false
   }
